Fix namespace URL
Due to this error, triples such as ?x :counter ?y were being inserted
with the wrong namespace. Hence, when they were queried by a system like
OntoManager, these triples could not be retrieved (as OntoManager was
querying them with the correct namespace).

This problem was discovered when the ordering of query results went bad
in OntoManager, due to the "wrong" ?x :counter ?y triples that were
inserted in the jsonld output files of Ontoscript.